Prepare new structure for tests
