+++ .gitignore
@@ -106,6 +106,6 @@ ENV/
 .idea/
 
 # Test files
-tests/media
-tests/static
-tests/db.sqlite3
+test_project/media
+test_project/static
+test_project/db.sqlite3
