+++ pyproject.toml
@@ -258,7 +258,7 @@ ignore = [
   # https://docs.astral.sh/ruff/rules/line-too-long
   "E501",
 ]
-"**/tests/*" = [
+"**/test_project/*" = [
   # https://docs.astral.sh/ruff/rules/assert
   "S101",
   # https://docs.astral.sh/ruff/rules/hardcoded-password-func-arg
@@ -385,11 +385,11 @@ exclude = [
 ]
 
 [[tool.mypy.overrides]]
-module = "tests.*"
+module = "test_project.*"
 warn_unreachable = false
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "tests.settings"
+DJANGO_SETTINGS_MODULE = "test_project.settings"
 python_files = [
   "tests.py",
   "test_*.py",
