+++ test_project/fake_app/apps.py
@@ -4,5 +4,5 @@ from django.apps import AppConfig
 class IOExtensionsAppConfig(AppConfig):
     """Fake app config."""
 
-    name = "tests.fake_app"
+    name = "test_project.fake_app"
     verbose_name = "Import Export Fake App"
