+++ test_project/fake_app/factories.py
@@ -55,7 +55,7 @@ class ArtistImportJobFactory(factory.django.DjangoModelFactory):
         ArtistImportJobFactory(artists=[artist1, artist2])
     """
 
-    resource_path = "tests.fake_app.resources.SimpleArtistResource"
+    resource_path = "test_project.fake_app.resources.SimpleArtistResource"
     resource_kwargs: dict[str, str] = {}
 
     class Meta:
@@ -86,7 +86,7 @@ class ArtistImportJobFactory(factory.django.DjangoModelFactory):
 class ArtistExportJobFactory(factory.django.DjangoModelFactory):
     """Factory for creating ExportJob for Artist."""
 
-    resource_path = "tests.fake_app." "resources.SimpleArtistResource"
+    resource_path = "test_project.fake_app." "resources.SimpleArtistResource"
     resource_kwargs: dict[str, str] = {}
     file_format_path = "import_export.formats.base_formats.CSV"
