+++ test_project/tests/test_models/test_import/test_cancel_import_job.py
@@ -3,8 +3,8 @@ from pytest_mock import MockerFixture
 
 from import_export_extensions.models import ImportJob
 
-from ...fake_app.factories import ArtistImportJobFactory
-from ...fake_app.models import Artist
+from ....fake_app.factories import ArtistImportJobFactory
+from ....fake_app.models import Artist
 
 
 def test_cancel_import_error_status(
